# [MXNET-1249] Fix Object Detector Performance with GPU